refactor: rename frequency_response/sweep.py to orchestrate.py
The module was named sweep.py while also exporting a function named
sweep(), so the package's `sweep` attribute ended up bound to the
function, shadowing the sweep submodule. A bare `import
scpi_control.frequency_response.sweep` silently got the function
instead of the module, and a test needed an importlib.import_module
workaround to reach the real submodule for monkeypatching.

Rename to orchestrate.py (matching the module's own docstring) removes
the collision entirely; the exported sweep() function and the
documented `from scpi_control.frequency_response import sweep` form
are unchanged. Drops the now-unnecessary importlib workaround and the
warning comments that only existed to flag the shadowing.
